some webserver code refactoring
