+++ webserver/src/server.cpp
@@ -1,8 +1,6 @@
+#include <WiFi.h>
 #include "server.h"
 
-AsyncWebServer server(80);
-// Create an Event Source on /
-AsyncEventSource events("/");
 
 
 void setup(){
@@ -13,7 +11,7 @@ void setup(){
         return;
     }
 
-    WiFi.begin(ssid, password);
+    WiFi.begin(wifi_ssid, wifi_password);
 
     while (WiFi.status() != WL_CONNECTED) {
         delay(1000);
@@ -23,26 +21,18 @@ void setup(){
     Serial.print("Connection successful, server address: ");
     Serial.println(WiFi.localIP());
 
-    server.addHandler(&events);
+    serverAddSseHandler();
 
+    //loading webserver frontend assets
     loadFSHTML();
     loadFSCSS();
     loadFSJavaScript();
     loadFSStatic();
 
     // firmwareUpdate();
-    // Handle Web Server Events
-    events.onConnect([](AsyncEventSourceClient *client){
-    if(client->lastId()){
-        Serial.printf("Client reconnected! Last message ID that it got is: %u\n", client->lastId());
-    }
-
-    // send event with message "hello!", id current millis
-    // and set reconnect delay to 200ms
-    client->send("hello!", NULL, millis(), 200);
-    });
-
-    server.begin();
+    // Handle Web server Events
+    sseHandleEvents();
+    serverBeginServe();
 }
 
 void loop(){
@@ -55,14 +45,14 @@ void loop(){
         if (weight<=45) weight += flow;
         else weight = 0;
 
-        // Send Events to the Web Server with the Sensor Readings
-        events.send("ping",NULL,millis());
-        events.send(String(getSeconds()).c_str(),"seconds",millis());
-        events.send(String(getMinutes()).c_str(),"minutes",millis());
-        events.send(String(temp).c_str(),"temp",millis());
-        events.send(String(press).c_str(),"press",millis());
-        events.send(String(flow).c_str(),"flow",millis());
-        events.send(String(weight).c_str(),"weight",millis());
+        // Send Events to the Web server with the Sensor Readings
+        sseSend("ping",NULL);
+        sseSend(String(getSeconds()).c_str(),"seconds");
+        sseSend(String(getMinutes()).c_str(),"minutes");
+        sseSend(String(temp).c_str(),"temp");
+        sseSend(String(press).c_str(),"press");
+        sseSend(String(flow).c_str(),"flow");
+        sseSend(String(weight).c_str(),"weight");
 
         timer = millis();
     }
@@ -75,70 +65,4 @@ long getSeconds() {
 
 long getMinutes() {
     return millis() / 1000 / 60;
-}
-
-void loadFSJavaScript() {
-    //JAVASCRIPT-JS
-    server.on("/assets/js/bootstrap.min.js", HTTP_GET, [](AsyncWebServerRequest *request){
-        request->send(LittleFS, "/assets/js/bootstrap.min.js", "text/javascript");
-    });
-    server.on("/assets/js/chart.js", HTTP_GET, [](AsyncWebServerRequest *request){
-        request->send(LittleFS, "/assets/js/chart.js", "text/javascript");
-    });
-    server.on("/assets/js/custom.js", HTTP_GET, [](AsyncWebServerRequest *request){
-        request->send(LittleFS, "/assets/js/custom.js", "text/javascript");
-    });
-    server.on("/assets/js/easypiechart-data.js", HTTP_GET, [](AsyncWebServerRequest *request){
-        request->send(LittleFS, "/assets/js/easypiechart-data.js", "text/javascript");
-    });
-    server.on("/assets/js/easypiechart.js", HTTP_GET, [](AsyncWebServerRequest *request){
-        request->send(LittleFS, "/assets/js/easypiechart.js", "text/javascript");
-    });
-    server.on("/assets/js/jquery-1.10.2.js", HTTP_GET, [](AsyncWebServerRequest *request){
-        request->send(LittleFS, "/assets/js/jquery-1.10.2.js", "text/javascript");
-    });
-    server.on("/assets/js/uiuiui-scripts.js", HTTP_GET, [](AsyncWebServerRequest *request){
-        request->send(LittleFS, "/assets/js/uiuiui-scripts.js", "text/javascript");
-    });
-}
-
-void loadFSCSS() {
-    // CSS
-    server.on("/assets/css/bootstrap-theme.min.css", HTTP_GET, [](AsyncWebServerRequest *request){
-        request->send(LittleFS, "/assets/css/bootstrap-theme.min.css", "text/css");
-    });
-    server.on("/assets/css/bootstrap.css", HTTP_GET, [](AsyncWebServerRequest *request){
-        request->send(LittleFS, "/assets/css/bootstrap.css", "text/css");
-    });
-    server.on("/assets/css/custom-styles.css", HTTP_GET, [](AsyncWebServerRequest *request){
-        request->send(LittleFS, "/assets/css/custom-styles.css", "text/css");
-    });
-    server.on("/assets/css/font-awesome.css", HTTP_GET, [](AsyncWebServerRequest *request){
-        request->send(LittleFS, "/assets/css/font-awesome.css", "text/css");
-    });
-    server.on("/assets/css/open_sans.css", HTTP_GET, [](AsyncWebServerRequest *request){
-        request->send(LittleFS, "/assets/css/open_sans.css", "text/css");
-    });
-}
-
-void loadFSHTML() {
-    // HTML
-    // server.on("/", HTTP_GET, [](AsyncWebServerRequest *request){
-    //     request->send(LittleFS, "/index.html", "text/html");
-    // });
-    server.on("/index.html", HTTP_GET, [](AsyncWebServerRequest *request){
-        request->send(LittleFS, "/index.html", "text/html");
-    });
-}
-
-void loadFSStatic() {
-    server.on("/assets/img/GAGGIUINO_ICON_transp.png", HTTP_GET, [](AsyncWebServerRequest *request){
-            request->send(LittleFS, "/assets/img/GAGGIUINO_ICON_transp.png", "text/plain");
-    });
-    server.on("/assets/img/uiui-light.png", HTTP_GET, [](AsyncWebServerRequest *request){
-            request->send(LittleFS, "/assets/img/uiui-light.png", "text/plain");
-    });
-    server.on("/assets/img/uiui-dark.png", HTTP_GET, [](AsyncWebServerRequest *request){
-            request->send(LittleFS, "/assets/img/uiui-dark.png", "text/plain");
-    });
 }
